fix(characters): read abilities directly in server component
CharacterSkills was fetching the public API at the hardcoded prod URL
even at build time, so SSG pages depended on whichever version of the
data prod was currently serving. After the previous commit moved the
local route to the new generated dataset (430 entries), the production
API still returned the legacy 119 entries until deployment, leaving
roughly two thirds of the SSG character pages reporting "ページは存在
しません" in local dev and in any pre-deploy build.

Switch the server component to import the JSON directly and compute
last_updated from the file mtime, matching the route. With no internal
caller left, drop the now-unused /api/characterAbilities route.
